roles and login working i think

+++ frontend/app.py
@@ -1,7 +1,69 @@
 import requests
 import json
+import hashlib
+import os
+import sqlite3
+from getpass import getpass
 
 BASE_URL = "http://localhost:8000"
+
+def connect_db():
+    return sqlite3.connect('system.db')
+
+def hash_password(password, salt):
+    return hashlib.pbkdf2_hmac('sha256', password.encode(), salt.encode(), 100000).hex()
+
+def login():
+    db = connect_db()
+    #only for test next line
+    #print(hash_password('temp', 'random_salt'))
+    username = input("Enter username: ")
+    password = getpass("Enter password: ")
+    
+    try:
+        cursor = db.cursor()
+        cursor.execute("SELECT password_hash, salt, role FROM User WHERE username = ?", (username,))
+        user = cursor.fetchone()
+        db.close()
+        if user:
+            hashed_password = hash_password(password, user[1])
+
+            if hashed_password == user[0] == '4d997b5e8d99318eb2d0a062a8a7b5901e16afcc0db7c114ec5c8c9ad5d1b215':
+                print("You currently are using a temp password and need to change it: ")
+                global role
+                role = user[2]
+                change_user_password()
+                return username, user[2]
+            elif hashed_password == user[0]:
+                return username, user[2]  # Return the role of the user
+            else:
+                print("Incorrect password")
+        else:
+            print("User not found")
+    except Exception as e:
+        print(f"Database error: {str(e)}")
+        db.close()
+    return None, None
+
+def logout():
+    global username, role
+    username = None
+    role = None
+    print("Logged out successfully. Returning to login page.")
+
+''' def authenticated_session(func):
+    def wrapper(*args, **kwargs):
+        if 'role' not in globals():
+            global username, role
+            credentials = login()
+            if credentials:
+                username, role = credentials
+                return func(*args, **kwargs)
+            else:
+                print("Authentication failed. Please try again.")
+        else:
+            return func(*args, **kwargs)
+    return wrapper '''
 
 def get_machine_choice():
     """
@@ -20,16 +82,26 @@ def get_machine_choice():
         return
     return machine
 
+# @authenticated_session
 def make_reservation():
+    if role not in ['admin', 'scheduler', 'customer']:
+        print("You do not have permission to make reservations.")
+        return
+
     try:
-        customer_name = input("\nEnter customer name: ")
+        # customer_name = input("\nEnter customer name: ")
+        if role == 'customer':
+            customer_name = username  # Customers can only make reservations for themselves
+        else:
+            customer_name = input("\nEnter customer name: ")  # Schedulers can make reservations for anyone
+        
         machine = get_machine_choice()
         if not machine:
             return    
         start_date = input("\nEnter start date (YYYY-MM-DD HH:MM): ")
         end_date = input("\nEnter end date (YYYY-MM-DD HH:MM): ")
 
-        data ={
+        data = {
             "customer_name": customer_name,
             "machine_name": machine,
             "start_date": start_date,
@@ -45,10 +117,31 @@ def make_reservation():
     except Exception as e:
         print(f"Error occurred while adding reservation: {str(e)}")
 
+# @authenticated_session
 def cancel_reservation():
+    if role not in ['admin', 'scheduler', 'customer']:
+        print("You do not have permission to cancel reservations.")
+        return
+
     try:
         reservation_id = input("\nEnter id of the reservation to cancel: ")
-        response = requests.delete(f"{BASE_URL}/reservations/{reservation_id}")
+        if role == 'customer':
+            db = connect_db()
+            cursor = db.cursor()
+            if role == 'customer':
+                # Verify that the reservation belongs to the logged-in customer
+                cursor.execute("SELECT customer FROM Reservation WHERE reservation_id = ?", (reservation_id,))
+                reservation = cursor.fetchone()
+                if reservation is None:
+                    print("Reservation not found.")
+                    return
+                elif reservation[0] != username: # assume customer in reservation is username of user
+                    print("You can only cancel your own reservations.")
+                    return
+
+            response = requests.delete(f"{BASE_URL}/reservations/{reservation_id}") # , json={'username': username}
+        else:
+            response = requests.delete(f"{BASE_URL}/reservations/{reservation_id}")
         if response.status_code == 200:
             refund = response.json()['refund']
             print("\nReservation cancelled successfully\nRefund amount: ", refund)
@@ -57,50 +150,99 @@ def cancel_reservation():
     except Exception as e:
         print("An error occurred ", str(e))
 
-
+# @authenticated_session
 def list_reservations():
-    print("\n\n1. List reservations for a given date range\n")
-    print("2. List the reservations for a given machine for a given date range\n")
-    print("3. List the reservations for a given customer for a given date range\n\n")
-    choice = input("\nEnter your choice: ")
-    start_date = input("\nEnter start date (YYYY-MM-DD HH:MM)  :")
-    end_date = input("\nEnter end date (YYYY-MM-DD HH:MM)  :")
-    params = {
-        "start": start_date,
-        "end": end_date
-    } # url paramaters
+    if role not in ['admin', 'scheduler', 'customer']:
+        print("You do not have permission to view reservations.")
+        return
 
-    try: 
-        if choice == '1':
-            response = requests.get(f"{BASE_URL}/reservations", params=params)
+    if role in ['admin', 'scheduler']:
+        print("\n\n1. List reservations for a given date range\n")
+        print("2. List the reservations for a given machine for a given date range\n")
+        print("3. List the reservations for a given customer for a given date range\n\n")
+        choice = input("\nEnter your choice: ")
+        start_date = input("\nEnter start date (YYYY-MM-DD HH:MM)  :")
+        end_date = input("\nEnter end date (YYYY-MM-DD HH:MM)  :")
+        params = {
+            "start": start_date,
+            "end": end_date
+        } # url parameters
 
-        elif choice == '2':
-            machine = get_machine_choice()
-            if not machine:
-                return    
-            response = requests.get(f"{BASE_URL}/reservations/machines/{machine}", params=params)
+        try:
+            if choice == '1':
+                response = requests.get(f"{BASE_URL}/reservations", params=params) # !!!! no functionality for only one customer
 
-        elif choice == '3':
-            customer = input("\nEnter customer name: ")
-            response = requests.get(f"{BASE_URL}/reservations/customers/{customer}", params=params)
+            elif choice == '2':
+                machine = get_machine_choice()
+                if not machine:
+                    return    
+                response = requests.get(f"{BASE_URL}/reservations/machines/{machine}", params=params) # !!!! no functionality for only one customer
 
-        else:
-            print("Invalid choice. Back to the main menu")
-            return
+            elif choice == '3':
+                #if role == 'customer':
+                #    customer = username  # Customers can only make reservations for themselves
+                #else:
+                customer = input("\nEnter customer name: ")  # Schedulers can make reservations for anyone
+                response = requests.get(f"{BASE_URL}/reservations/customers/{customer}", params=params)
 
-        if response.status_code == 200:
-            reservations_data = response.json()
-            if 'reservations' in reservations_data:
-                data = reservations_data['reservations']
-                print(json.dumps(data, indent=2))
-            else: # no reservations found
-                print(response.json()['message'])
+            else:
+                print("Invalid choice. Back to the main menu")
+                return
 
-        else:
-            print(response.json()['detail'])
+            if response.status_code == 200:
+                reservations_data = response.json()
+                if 'reservations' in reservations_data:
+                    data = reservations_data['reservations']
+                    print(json.dumps(data, indent=2))
+                else: # no reservations found
+                    print(response.json()['message'])
 
-    except Exception as e:
-        print("Error occurred in retrieving reservations ",str(e))
+            else:
+                print(response.json()['detail'])
+
+        except Exception as e:
+            print("Error occurred in retrieving reservations ", str(e))
+    else:
+        # in the customer case
+        print("\n\n1. List reservations for a given date range\n")
+        print("2. List the reservations for a given machine for a given date range\n")
+        choice = input("\nEnter your choice: ")
+        start_date = input("\nEnter start date (YYYY-MM-DD HH:MM)  :")
+        end_date = input("\nEnter end date (YYYY-MM-DD HH:MM)  :")
+        params = {
+            "start": start_date,
+            "end": end_date
+        } # url parameters
+
+        try:
+            customer = username
+            if choice == '1':
+                response = requests.get(f"{BASE_URL}/reservations/customers/{customer}", params=params)
+
+            elif choice == '2':
+                machine = get_machine_choice()
+                if not machine:
+                    return    
+                response = requests.get(f"{BASE_URL}/reservations/machines/{machine}/customers/{customer}", params=params)
+
+            else:
+                print("Invalid choice. Back to the main menu")
+                return
+
+            if response.status_code == 200:
+                reservations_data = response.json()
+                if 'reservations' in reservations_data:
+                    data = reservations_data['reservations']
+                    print(json.dumps(data, indent=2))
+                else: # no reservations found
+                    print(response.json()['message'])
+
+            else:
+                print(response.json()['detail'])
+
+        except Exception as e:
+            print("Error occurred in retrieving reservations ", str(e))
+
 
 def exit_system():
     try:
@@ -118,23 +260,200 @@ def exit_system():
     except Exception as e:
         print("Error while saving changes ",str(e))
 
-def main():
-    while True:
-        print("\nWelcome to the Reservation System!\n")
-        print("\n\n1. Make reservation\n2. Cancel reservation\n3. List reservations\n4. Quit\n\n")
-        choice = input("\nEnter your choice: ")
-        if choice == '1':
-            make_reservation()
-        elif choice == '2':
-            cancel_reservation()
-        elif choice == '3':
-            list_reservations()
-        elif choice == '4':
-            exit_system()
-            break
-        else:
-            print("Invalid choice, please try again")
 
+# admin management
+
+# @authenticated_session
+def add_user():
+    if role != 'admin':
+        print("Unauthorized access. Only admins can add users.")
+        return
+    
+    username = input("Enter new username: ")
+    #temporary_password = input("Enter temporary password: ")
+    user_role = input("Enter user role (customer, scheduler, admin): ")
+    #salt = os.urandom(32).hex()
+    password_hash = hash_password('temp', 'random_salt')
+    
+    db = connect_db()
+    try:
+        cursor = db.cursor()
+        cursor.execute("INSERT INTO User (username, password_hash, salt, role) VALUES (?, ?, ?, ?)", 
+                       (username, password_hash, 'random_salt', user_role))
+        db.commit()
+        print("User added successfully with a temporary password.")
+    except Exception as e:
+        print("Failed to add user: ", str(e))
+    finally:
+        db.close()
+
+# @authenticated_session
+def remove_user():
+    if role != 'admin':
+        print("Unauthorized access. Only admins can remove users.")
+        return
+    
+    username_to_remove = input("Enter username of the user to remove: ")
+    db = connect_db()
+    try:
+        cursor = db.cursor()
+        # Prevent admin from removing the last admin
+        cursor.execute("SELECT COUNT(*) FROM User WHERE role = 'admin'")
+        count = cursor.fetchone()[0]
+        cursor.execute("SELECT role FROM User WHERE username = ?", (username_to_remove,))
+        user_role = cursor.fetchone()[0]
+        if count <= 1 and user_role == 'admin':
+            print("Cannot remove the last admin.")
+            return
+        
+        cursor.execute("DELETE FROM User WHERE username = ?", (username_to_remove,))
+        db.commit()
+        print("User removed successfully.")
+    except Exception as e:
+        print("Failed to remove user: ", str(e))
+    finally:
+        db.close()
+
+# @authenticated_session
+def change_user_role():
+    if role != 'admin':
+        print("Unauthorized access. Only admins can change user roles.")
+        return
+    
+    username_to_change = input("Enter username of the user to change role: ")
+    new_role = input("Enter new role (customer, scheduler, admin): ")
+    db = connect_db()
+    try:
+        cursor = db.cursor()
+
+        cursor.execute("SELECT COUNT(*) FROM User WHERE role = 'admin'")
+        count = cursor.fetchone()[0]
+        cursor.execute("SELECT role FROM User WHERE username = ?", (username_to_change,))
+        user_role = cursor.fetchone()[0]
+        if count <= 1 and user_role == 'admin':
+            print("Cannot change role of the last admin.")
+            return
+
+        cursor.execute("UPDATE User SET role = ? WHERE username = ?", (new_role, username_to_change))
+        db.commit()
+        print("User role updated successfully.")
+    except Exception as e:
+        print("Failed to change user role: ", str(e))
+    finally:
+        db.close()
+
+def change_user_password():
+    db = connect_db()
+    try:
+        '''current_password = getpass("Enter your current password: ")
+        cursor = db.cursor()
+        cursor.execute("SELECT password_hash, salt FROM User WHERE username = ?", (username,))
+        user = cursor.fetchone()
+        if user is None:
+            print("User not found.")
+            return
+        
+        # Verify the current password
+        hashed_password = hash_password(current_password, user[1])
+        if hashed_password != user[0]:
+            print("Incorrect password.")
+            return '''
+        
+        cursor = db.cursor()
+
+        # Enter new password
+        new_password = getpass("Enter your new password: ")
+        new_password_confirm = getpass("Confirm your new password: ")
+        if new_password != new_password_confirm:
+            print("Passwords do not match.")
+            return
+        #print(role)
+        # Hash the new password
+        new_salt = os.urandom(32).hex()  # Generate a new salt
+        new_hashed_password = hash_password(new_password, new_salt)
+        
+        if role == 'customer':
+            customer_name = username  # Customers can only make reservations for themselves
+        else:
+            customer_name = input("\nEnter customer name to change: ")  # Schedulers can make reservations for anyone
+
+        # Update the password in the database
+        cursor.execute("UPDATE User SET password_hash = ?, salt = ? WHERE username = ?", 
+                       (new_hashed_password, new_salt, customer_name))
+        db.commit()
+        print("Password changed successfully.")
+        
+    except Exception as e:
+        print(f"Error while changing password: {str(e)}")
+    finally:
+        db.close()
+
+
+
+
+
+def show_menu_by_role(role):
+    actions = {
+        'customer': ["1. Make reservation", "2. Cancel reservation", "3. List reservations", "4. Change Password", "5. Logout", "6. Exit"],
+        'scheduler': ["1. Make reservation", "2. Cancel reservation", "3. List reservations", "4. Change Password", "5. Logout", "6. Exit"],
+        'admin': ["1. Make reservation", "2. Cancel reservation", "3. List reservations", "4. Change Password", "5. Logout", "6. Exit", "7. Add user", "8. Remove user", "9. Change user role"]
+    }
+    if role in actions:
+        return actions[role]
+    return []
+
+def main():
+    print("\nWelcome to the Reservation System!\n")
+    global username, role
+    while True:
+        username, role = login()
+        print("\n")
+        break_outer = False
+        if username is not None:
+            while True:
+                options = show_menu_by_role(role)
+                if not options:
+                    print("No available actions for this role.")
+                    return
+                for option in options:
+                    print(option)
+                
+                choice = input("\nEnter your choice: ")
+                if choice == '1':
+                    #if role in ['customer', 'scheduler']:
+                    make_reservation()
+                    #elif role == 'admin':
+                    #    add_user()
+                elif choice == '2':
+                    #if role in ['customer', 'scheduler']:
+                    cancel_reservation()
+                    #elif role == 'admin':
+                    #    remove_user()
+                elif choice == '3':
+                    #if role in ['customer', 'scheduler']:
+                    list_reservations()
+                    #elif role == 'admin':
+                    #    change_user_role()
+                elif choice == '4':
+                    change_user_password()
+                elif choice == '5':
+                    logout()
+                    break
+                elif choice == '6':
+                    exit_system()
+                    break_outer = True
+                    break
+                elif choice == '7' and role == 'admin':
+                    add_user()
+                elif choice == '8' and role == 'admin':
+                    remove_user()
+                elif choice == '9' and role == 'admin':
+                    change_user_role()
+                else:
+                    print("Invalid choice, please try again")
+        if break_outer:
+            break
+        
 
 if __name__ == "__main__":
     main()
